+++ packages/components/text-field/src/TextField.test.tsx
@@ -1,0 +1,14 @@
+import { render, screen } from '@testing-library/react'
+import { describe, expect, it } from 'vitest'
+
+import { TextField } from '.'
+
+describe('TextField', () => {
+  it('should render a label as children', () => {
+    const label = 'Title'
+
+    render(<TextField>{label}</TextField>)
+
+    expect(screen.getByLabelText(label)).toBeVisible()
+  })
+})